+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/tools/cli/commands/toggle/AutoRecoveryCommand.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.bookkeeper.tools.cli.commands.autorecovery;
+package org.apache.bookkeeper.tools.cli.commands.toggle;
 
 import com.beust.jcommander.Parameter;
 import com.google.common.util.concurrent.UncheckedExecutionException;
