+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/tools/cli/commands/toggle/package-info.java
@@ -20,4 +20,4 @@
 /**
  * This package provides all autorecovery commands.
  */
-package org.apache.bookkeeper.tools.cli.commands.autorecovery;
+package org.apache.bookkeeper.tools.cli.commands.toggle;
